+++ packages/ts-logkit/src/transports/index.ts
@@ -1,0 +1,2 @@
+export { createConsoleTransport } from "./console";
+export { combineTransports } from "./utils";