+++ packages/ts-logkit/CHANGELOG.md
@@ -22,11 +22,13 @@
 
 ### Added
 
+- `src/transports:combineTransports` function for consumers to conveniently combine multiple transports
 -
 
 ### Changed
 
 - Simplified error handling: removed any verbose error or error verbosity features to keep the API simple and put the responsibility of what to log on the user
+- refactored `src/types`
 
 ### Breaking